@@ -8,7 +8,7 @@ style = "https://raw.githubusercontent.com/wemake-services/wemake-python-stylegu
 
 
 [tool.black]
-target-version = ['py36', 'py37', 'py38']
+target-version = ['py37', 'py38', 'py39', 'py310']
 skip-string-normalization = true
 include = '\.pyi?$'
 
@@ -16,7 +16,7 @@ include = '\.pyi?$'
 [tool.poetry]
 name = "django-tenant-users"
 description = "A Django app to extend django-tenants to incorporate global multi-tenant users"
-version = "0.3.12"
+version = "1.0.0"
 license = "MIT License"
 
 packages = [
@@ -39,51 +39,54 @@ keywords = [
 classifiers = [
   "Environment :: Web Environment",
   "Framework :: Django",
-  "Framework :: Django :: 1.11",
+  "Framework :: Django :: 2.2",
+  "Framework :: Django :: 3.2",
+  "Framework :: Django :: 4.0",
   "Intended Audience :: Developers",
   "License :: OSI Approved :: MIT License",
   "Operating System :: OS Independent",
   "Programming Language :: Python",
   "Programming Language :: Python :: 3",
-  "Programming Language :: Python :: 3.5",
-  "Programming Language :: Python :: 3.6",
+  "Programming Language :: Python :: 3.7",
+  "Programming Language :: Python :: 3.8",
+  "Programming Language :: Python :: 3.9",
+  "Programming Language :: Python :: 3.10",
   "Topic :: Internet :: WWW/HTTP",
   "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
 ]
 
 [tool.poetry.dependencies]
-python = "^3.6.2"
+python = "^3.7"
 
 
 [tool.poetry.dev-dependencies]
 django = "^3.2"
+django-tenants = "^3.4.2"
 
-mypy = "^0.902"
-django-stubs = "^1.8"
+mypy = "^0.942"
+django-stubs = "^1.10"
 
-wemake-python-styleguide = "^0.15"
-flake8-pytest-style = "^1.4"
-nitpick = "^0.26"
+wemake-python-styleguide = "^0.16"
+flake8-pytest-style = "^1.6"
+nitpick = "^0.32"
 
 safety = "^1.10"
 
 pytest = "^6.2"
-pytest-cov = "^2.12"
-pytest-randomly = "^3.0"
+pytest-cov = "^3.0"
+pytest-randomly = "^3.11"
 
-sphinx = "^4.0"
-sphinx-autodoc-typehints = "^1.12"
-doc8 = "^0.8"
-m2r2 = "^0.2"
-tomlkit = "^0.7"
-pytest-pythonpath = "^0.7.3"
-pytest-django = "^4.4.0"
-black = "^21.6b0"
-python-decouple = "^3.4"
+sphinx = "^4.5"
+sphinx-autodoc-typehints = "^1.18"
+doc8 = "^0.11"
+m2r2 = "^0.3"
+tomlkit = "^0.10"
+pytest-pythonpath = "^0.7.4"
+pytest-django = "^4.5.2"
+black = "^22.3"
+python-decouple = "^3.6"
 psycopg2-binary = "~2.8.0"
 pytest-deadfixtures = "^2.2.1"
-pytest-testmon = "^1.1.1"
-pytest-timeout = "^1.4.2"
-hypothesis = "^6.14.3"
-nox = "^2021.6.12"
-nox-poetry = "^0.8.6"
+pytest-testmon = "^1.3.1"
+pytest-timeout = "^2.1.0"
+hypothesis = "^6.43.3"
