@@ -2,29 +2,31 @@ import time
 
 from django.conf import settings
 from django.contrib.auth import get_user_model
+from django.contrib.auth.hashers import make_password
 from django.contrib.auth.models import AbstractBaseUser, BaseUserManager
 from django.db import connection, models
 from django.dispatch import Signal
 from django.utils import timezone
-from django.utils.translation import ugettext_lazy as _
+from django.utils.translation import gettext_lazy as _
+from django_tenants.models import TenantMixin
+from django_tenants.utils import get_public_schema_name, get_tenant_model
 
-from tenant_users.compat import (
-    TenantMixin,
-    get_public_schema_name,
-    get_tenant_model,
+from tenant_users.permissions.models import (
+    PermissionsMixinFacade,
+    UserTenantPermissions,
 )
 
 # An existing user removed from a tenant
-tenant_user_removed = Signal(providing_args=['user', 'tenant'])
+tenant_user_removed = Signal()
 
 # An existing user added to a tenant
-tenant_user_added = Signal(providing_args=['user', 'tenant'])
+tenant_user_added = Signal()
 
 # A new user is created
-tenant_user_created = Signal(providing_args=['user'])
+tenant_user_created = Signal()
 
 # An existing user is deleted
-tenant_user_deleted = Signal(providing_args=['user'])
+tenant_user_deleted = Signal()
 
 
 class InactiveError(Exception):
@@ -232,11 +234,6 @@ class UserProfileManager(BaseUserManager):
         if not email:
             raise ValueError('Users must have an email address.')
 
-        # If no password is submitted, just assign a random one to lock down
-        # the account a little bit.
-        if not password:
-            password = self.make_random_password(length=30)
-
         email = self.normalize_email(email)
 
         profile = UserModel.objects.filter(email=email).first()
@@ -279,13 +276,18 @@ class UserProfileManager(BaseUserManager):
         is_staff=False,
         **extra_fields,
     ):
-        return self._create_user(
+        user = self._create_user(
             email,
             password,
             is_staff,
             False,
             **extra_fields,
         )
+
+        if not password:
+            user.set_unusable_password()
+
+        return user
 
     def create_superuser(self, password, email=None, **extra_fields):
         return self._create_user(
