@@ -2,14 +2,13 @@ import time
 
 from django.conf import settings
 from django.contrib.auth import get_user_model
-
-from tenant_users.compat import (
-    TENANT_SCHEMAS,
+from django_tenants.utils import (
     get_public_schema_name,
     get_tenant_domain_model,
     get_tenant_model,
     schema_context,
 )
+
 from tenant_users.tenants.models import ExistsError, InactiveError
 
 
@@ -23,20 +22,19 @@ def provision_tenant(tenant_name, tenant_slug, user_email):
 
     UserModel = get_user_model()
     TenantModel = get_tenant_model()
-    DomainModel = get_tenant_domain_model()
 
     user = UserModel.objects.get(email=user_email)
     if not user.is_active:
         raise InactiveError('Inactive user passed to provision tenant')
 
-    tenant_domain = '{0}.{1}'.format(tenant_slug, settings.TENANT_USERS_DOMAIN)
-
-    if TENANT_SCHEMAS:
-        if TenantModel.objects.filter(domain_url=tenant_domain).exists():
-            raise ExistsError('Tenant URL already exists')
+    if hasattr(settings, 'TENANT_SUBFOLDER_PREFIX'):
+        tenant_domain = tenant_slug
     else:
-        if DomainModel.objects.filter(domain=tenant_domain).exists():
-            raise ExistsError('Tenant URL already exists.')
+        tenant_domain = '{0}.{1}'.format(tenant_slug, settings.TENANT_USERS_DOMAIN)
+
+    DomainModel = get_tenant_domain_model()
+    if DomainModel.objects.filter(domain=tenant_domain).exists():
+        raise ExistsError('Tenant URL already exists.')
 
     time_string = str(int(time.time()))
     # Must be valid postgres schema characters see:
@@ -51,30 +49,19 @@ def provision_tenant(tenant_name, tenant_slug, user_email):
         # Wrap it in public schema context so schema consistency is maintained
         # if any error occurs
         with schema_context(get_public_schema_name()):
-            if TENANT_SCHEMAS:
-                # Create a TenantModel object and tenant schema
-                tenant = TenantModel.objects.create(
-                    name=tenant_name,
-                    slug=tenant_slug,
-                    domain_url=tenant_domain,
-                    schema_name=schema_name,
-                    owner=user,
-                )
+            tenant = TenantModel.objects.create(
+                name=tenant_name,
+                slug=tenant_slug,
+                schema_name=schema_name,
+                owner=user,
+            )
 
-            else:  # django-tenants
-                tenant = TenantModel.objects.create(
-                    name=tenant_name,
-                    slug=tenant_slug,
-                    schema_name=schema_name,
-                    owner=user,
-                )
-
-                # Add one or more domains for the tenant
-                domain = get_tenant_domain_model().objects.create(
-                    domain=tenant_domain,
-                    tenant=tenant,
-                    is_primary=True,
-                )
+            # Add one or more domains for the tenant
+            domain = get_tenant_domain_model().objects.create(
+                domain=tenant_domain,
+                tenant=tenant,
+                is_primary=True,
+            )
             # Add user as a superuser inside the tenant
             tenant.add_user(user)
     except:
