@@ -1,12 +1,11 @@
 from django.contrib.auth import get_user_model
 from django.db import connection
-
-from tenant_users.compat import (
-    TENANT_SCHEMAS,
+from django_tenants.utils import (
     get_public_schema_name,
     get_tenant_domain_model,
     get_tenant_model,
 )
+
 from tenant_users.tenants.models import ExistsError
 
 
@@ -38,48 +37,18 @@ def create_public_tenant(domain_url, owner_email, **owner_extra):
     profile.save()
 
     # Create public tenant
+    public_tenant = TenantModel.objects.create(
+        schema_name=public_schema_name,
+        name='Public Tenant',
+        owner=profile,
+    )
 
-    if TENANT_SCHEMAS:
-        public_tenant = TenantModel.objects.create(
-            domain_url=domain_url,
-            schema_name=public_schema_name,
-            name='Public Tenant',
-            owner=profile,
-        )
-    else:
-        public_tenant = TenantModel.objects.create(
-            schema_name=public_schema_name,
-            name='Public Tenant',
-            owner=profile,
-        )
-
-        # Add one or more domains for the tenant
-        get_tenant_domain_model().objects.create(
-            domain=domain_url,
-            tenant=public_tenant,
-            is_primary=True,
-        )
+    # Add one or more domains for the tenant
+    get_tenant_domain_model().objects.create(
+        domain=domain_url,
+        tenant=public_tenant,
+        is_primary=True,
+    )
 
     # Add system user to public tenant (no permissions)
     public_tenant.add_user(profile)
-
-
-def fix_tenant_urls(domain_url):
-    """
-    Helper function to update the domain urls on all tenants
-    Useful for domain changes in development
-    """
-    TenantModel = get_tenant_model()
-    public_schema_name = get_public_schema_name()
-
-    tenants = TenantModel.objects.all()
-    for tenant in tenants:
-        if tenant.schema_name == public_schema_name:
-            tenant.domain_url = domain_url
-        else:
-            # Assume the URL is wrong, parse out the subdomain
-            # and glue it back to the domain URL configured
-            slug = tenant.domain_url.split('.')[0]
-            new_url = '{0}.{1}'.format(slug, domain_url)
-            tenant.domain_url = new_url
-        tenant.save()
